+++ src/TextField/TextField.jsx
@@ -2,17 +2,23 @@ import FormControl from '@material-ui/core/FormControl'
 import FormHelperText from '@material-ui/core/FormHelperText'
 import InputBase from '@material-ui/core/InputBase'
 import PropTypes from 'prop-types'
-import React from 'react'
+import React, { useEffect, useState } from 'react'
 import isBoolean from 'lodash/isBoolean'
 import withStyles from '@material-ui/core/styles/withStyles'
 
 import FormLabel from '../form/FormLabel'
+import { Grid } from '../index'
 
 const styles = theme => {
   const light = theme.palette.type === 'light'
   const borderColor = light ? 'rgba(0, 0, 0, 0.42)' : 'rgba(255, 255, 255, 0.7)'
 
   return {
+    // This is a bit confusing when you first look at it, but what we're doing here
+    // is treating the subcomponent InputBase as the root component so that we can
+    // take advantage of it's named css rules.
+    //
+    // https://material-ui.com/api/input-base/#css
     root: {
       backgroundColor: theme.palette.common.white,
       border: `1px solid ${borderColor}`,
@@ -22,11 +28,6 @@ const styles = theme => {
         duration: theme.transitions.duration.shorter
       })
     },
-    formControl: {
-      'label + &': {
-        marginTop: 8
-      }
-    },
     focused: {
       borderColor: theme.palette.primary.main
     },
@@ -35,6 +36,14 @@ const styles = theme => {
     },
     input: {
       padding: '10px 12px'
+    },
+    // From here on these are custom styles for components which are not descendents
+    // of InputBase.
+    formLabel: {
+      marginBottom: 8
+    },
+    formHelperText: {
+      lineHeight: '1.5'
     }
   }
 }
@@ -54,8 +63,8 @@ const styles = theme => {
  * ```
  */
 export const TextField = ({
-  InputProps,
   autoComplete,
+  classes,
   disabled,
   error,
   fullWidth,
@@ -63,9 +72,47 @@ export const TextField = ({
   id,
   label,
   required,
+  maxLength,
+  maxLengthCountText,
   ...other
 }) => {
+  const [errorState, setErrorState] = useState(error)
   const helperTextId = helperText && id ? `${id}-helper-text` : undefined
+
+  // Max length validation with character count
+  const [characterCount, setCharacterCount] = useState(0)
+  const hasLengthValidation = maxLength > 0
+  const maxValidationHelperText = `${characterCount}/${maxLength} ${maxLengthCountText}`
+  const showHelperText = helperText || hasLengthValidation
+
+  useEffect(() => {
+    if (hasLengthValidation) {
+      const { value: valueProp, defaultValue: defaultValueProp } = other
+      const inputValue = valueProp || defaultValueProp
+
+      if (inputValue) { setCharacterCount(inputValue.length) }
+    }
+  }, [])
+
+  const onChange = (event) => {
+    // Call the `other.onChange` prop function, if any
+    if (other.onChange) { other.onChange(event) }
+
+    if (hasLengthValidation) {
+      // We don't want to miscount controlled inputs
+      const { value: valueProp } = other
+
+      if (valueProp) {
+        setCharacterCount(valueProp.length)
+        return
+      }
+
+      const currentCharacterCount = event.target.value.length
+
+      setCharacterCount(currentCharacterCount)
+      setErrorState(currentCharacterCount > maxLength)
+    }
+  }
 
   // Sugar for setting autoComplete using a boolean value.
   if (isBoolean(autoComplete)) {
@@ -73,18 +120,25 @@ export const TextField = ({
   }
 
   return (
-    <FormControl aria-describedby={helperTextId} disabled={disabled} error={error} fullWidth={fullWidth}>
+    <FormControl aria-describedby={helperTextId} disabled={disabled} error={errorState} fullWidth={fullWidth}>
       {label && (
-        <FormLabel htmlFor={id} required={required}>
+        <FormLabel className={classes.formLabel} htmlFor={id} required={required}>
           {label}
         </FormLabel>
       )}
 
-      <InputBase {...other} autoComplete={autoComplete} id={id} {...InputProps} />
+      <InputBase {...other} classes={classes} autoComplete={autoComplete} id={id} onChange={onChange} />
 
-      {helperText && (
-        <FormHelperText id={helperTextId}>
-          {helperText}
+      {showHelperText && (
+        <FormHelperText id={helperTextId} className={classes.formHelperText}>
+          <Grid component='span' justify='space-between' container spacing={2}>
+            <Grid component='span' item xs>{helperText}</Grid>
+            {hasLengthValidation && (
+              <Grid component='span' item xs style={{ textAlign: 'right' }}>
+                {maxValidationHelperText}
+              </Grid>
+            )}
+          </Grid>
         </FormHelperText>
       )}
     </FormControl>
@@ -164,6 +218,19 @@ TextField.propTypes = {
   required: PropTypes.bool,
 
   /**
+   * The maximum number of characters allowed in the text field before it's marked as invalid.
+   * Defaults to 0.
+   *
+   * Any value above 0 will add a character counter to the text field helper text
+   */
+  maxLength: PropTypes.number,
+
+  /**
+   * Text to be placed at the right side of character count. Example: `5/10 characters`
+   */
+  maxLengthCountText: PropTypes.string,
+
+  /**
    * The type of the text field.
    */
   type: PropTypes.oneOf(['email', 'text', 'password', 'number']),
@@ -179,6 +246,8 @@ TextField.defaultProps = {
   disabled: false,
   error: false,
   fullWidth: false,
+  maxLength: 0,
+  maxLengthCountText: '',
   required: false,
   type: 'text'
 }